# project/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: AI shot API
  description: 海外 AI 短剧平台后端 API 接口规范
  version: "1.0.0"
  contact:
    name: AI shot Team
  license:
    name: Proprietary
servers:
  - url: https://api.venaitry.com/api/v1
    description: Production server
  - url: https://staging-api.venaitry.com/api/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access Token (有效期15分钟)
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          description: 0=成功, 非0=错误码
          example: 0
        message:
          type: string
          example: "success"
        data:
          type: object
          nullable: true
        meta:
          type: object
          nullable: true
          properties:
            page:
              type: integer
              example: 1
            page_size:
              type: integer
              example: 20
            total:
              type: integer
              example: 150
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - nickname
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
        nickname:
          type: string
          minLength: 1
          maxLength: 100
...
    AuthTokens:
      type: object
      properties:
        access_token:
          type: string
          description: JWT Access Token (15min 有效期)
        refresh_token:
          type: string
          description: Refresh Token (30天 有效期)
        expires_in:
          type: integer
          description: Access Token 过期秒数
          example: 900
        user:
          $ref: '#/components/schemas/User'
...
paths:
...
  /auth/register:
    post:
      tags: [认证]
      summary: 邮箱注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        "201":
          description: 注册成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/AuthTokens'
        "409":
          description: 邮箱已存在
        "400":
          description: 参数校验失败
  /auth/login:
    post:
      tags: [认证]
      summary: 邮箱密码登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/AuthTokens'
        "401":
          description: 邮箱或密码错误
  /auth/oauth/{provider}:
    get:
      tags: [认证]
      summary: OAuth 登录跳转
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, apple, facebook]
      responses:
        "302":
          description: 重定向到 OAuth 提供商
  /auth/oauth/{provider}/callback:
    post:
      tags: [认证]
      summary: OAuth 回调
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, apple, facebook]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
                  description: OAuth 授权码
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/AuthTokens'
  /auth/refresh:
    post:
      tags: [认证]
      summary: 刷新 Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        "200":
          description: 刷新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/AuthTokens'
  /auth/logout:
    post:
      tags: [认证]
      summary: 登出
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 登出成功

  # ========== 用户资料 ==========
  /users/me:
    get:
      tags: [用户]
      summary: 获取当前用户资料
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 用户资料
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
    put:
      tags: [用户]
      summary: 更新当前用户资料
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        "200":
          description: 更新后的用户资料
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'

  # ========== 短剧内容 ==========
  /dramas:
    get:
      tags: [短剧]
      summary: 短剧列表（含推荐/分类筛选）
      parameters:
        - name: category_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, reviewing, archived]
          description: 管理后台按状态筛选
        - name: sort
          in: query
          schema:
            type: string
            enum: [latest, popular, trending]
            default: latest
        - name: keyword
          in: query
          schema:
            type: string
        - name: tags
          in: query
          schema:
            type: string
          description: 逗号分隔
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: 短剧列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Drama'
    post:
      tags: [短剧]
      summary: 创建短剧
      description: 管理员/创作者创建短剧草稿
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDramaRequest'
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Drama'
        "400":
          description: 参数校验失败
        "401":
          description: 未登录
        "403":
          description: 无权限（非管理员/创作者）
  /dramas/{id}:
    get:
      tags: [短剧]
      summary: 短剧详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 短剧详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Drama'
        "404":
          description: 短剧不存在
    put:
      tags: [短剧]
      summary: 更新短剧
      description: 管理员/创作者更新短剧元数据
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDramaRequest'
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Drama'
        "404":
          description: 短剧不存在
    delete:
      tags: [短剧]
      summary: 删除短剧
      description: 管理员删除短剧及其所有剧集
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 删除成功
        "404":
          description: 短剧不存在
  /dramas/{id}/status:
    put:
      tags: [短剧]
      summary: 更新短剧状态
      description: 管理员审核短剧（draft → reviewing → published/rejected）
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [draft, published, reviewing, archived]
      responses:
        "200":
          description: 状态更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Drama'
  /dramas/{id}/episodes:
    get:
      tags: [短剧]
      summary: 剧集列表
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            description: 短剧 ID
      responses:
        "200":
          description: 剧集列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Episode'
    post:
      tags: [短剧]
      summary: 添加剧集
      description: 管理员/创作者为短剧添加剧集（上传视频完成后调用）
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            description: 短剧 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEpisodeRequest'
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Episode'
  /episodes/{id}:
    get:
      tags: [短剧]
      summary: 剧集详情含播放地址
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 剧集详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Episode'
    put:
      tags: [短剧]
      summary: 更新剧集
      description: 管理员/创作者更新剧集元数据
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEpisodeRequest'
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Episode'
        "404":
          description: 剧集不存在
    delete:
      tags: [短剧]
      summary: 删除剧集
      description: 管理员删除剧集及其关联数据
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 删除成功
        "404":
          description: 剧集不存在
  /episodes/{id}/play:
    get:
      tags: [短剧]
      summary: 获取播放 URL（签名）
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: quality
          in: query
          schema:
            type: string
            enum: [1080p, 720p, 480p]
            default: 720p
        - name: language
          in: query
          schema:
            type: string
            description: 语言代码，不传则返回原始音频
      responses:
        "200":
          description: 播放信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/EpisodePlayInfo'

  # ========== 视频上传 ==========
  /videos/upload-url:
    post:
      tags: [视频上传]
      summary: 获取视频上传 URL
      description: 管理员/创作者请求上传视频，返回 S3/MinIO 预签名 URL
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        "200":
          description: 上传 URL
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/UploadUrlResponse'
        "400":
          description: 参数校验失败（格式不支持或文件过大）
  /videos/upload/complete:
    post:
      tags: [视频上传]
      summary: 上传完成回调
      description: 前端上传完成后通知后端，后端写入 video_assets 并创建 AI 任务
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadCompleteRequest'
      responses:
        "200":
          description: 上传完成
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/UploadCompleteResponse'
        "400":
          description: 参数错误

  # ========== 通用文件上传 ==========
  /files/upload-url:
    post:
      tags: [文件上传]
      summary: 获取通用文件上传 URL
      description: 返回 S3/MinIO 预签名 PUT URL，用于上传头像、封面等非视频文件。上传完成后使用 download_url 访问文件。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileUploadUrlRequest'
      responses:
        "200":
          description: 预签名上传 URL
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FileUploadUrlResponse'
        "400":
          description: 参数错误

  # ========== 收藏 ==========
  /favorites:
    post:
      tags: [收藏]
      summary: 添加收藏
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddFavoriteRequest'
      responses:
        "201":
          description: 收藏成功
        "409":
          description: 已收藏
    get:
      tags: [收藏]
      summary: 收藏列表
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: 收藏列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Favorite'
  /favorites/{id}:
    delete:
      tags: [收藏]
      summary: 取消收藏
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 取消收藏成功
        "404":
          description: 收藏不存在

  # ========== 评论 ==========
  /dramas/{id}/comments:
    post:
      tags: [评论]
      summary: 发布评论
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            description: 短剧 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        "201":
          description: 评论成功
    get:
      tags: [评论]
      summary: 评论列表
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            description: 短剧 ID
        - name: sort
          in: query
          schema:
            type: string
            enum: [latest, hottest]
            default: latest
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: 评论列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Comment'
  /comments/{id}:
    delete:
      tags: [评论]
      summary: 删除评论
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 删除成功
        "403":
          description: 无权限
  /comments/{id}/like:
    post:
      tags: [评论]
      summary: 点赞评论
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 操作成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        type: object
                        properties:
                          is_liked:
                            type: boolean
                          likes_count:
                            type: integer

  # ========== 支付订阅 ==========
...
  /subscriptions/create:
    post:
      tags: [订阅]
      summary: 创建订阅
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        "200":
          description: 创建成功，返回支付链接
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        type: object
                        properties:
                          subscription_id:
                            type: integer
                          payment_url:
                            type: string
                            format: uri
  /subscriptions/cancel:
    post:
      tags: [订阅]
      summary: 取消订阅
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 取消成功
  /subscriptions/status:
    get:
      tags: [订阅]
      summary: 订阅状态
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 订阅状态
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Subscription'
  /payments/webhook/{channel}:
    post:
      tags: [支付]
      summary: 支付回调 Webhook
      parameters:
        - name: channel
          in: path
          required: true
          schema:
            type: string
            enum: [paypal, stripe, apple, google]
      requestBody:
        content:
          application/json: {}
      responses:
        "200":
          description: 回调处理成功

  # ========== AI Internal（AI Worker → 后端回调）==========
  /internal/ai/jobs/{job_id}/status:
    put:
      tags: [AI Internal]
      summary: 更新 AI 任务状态
      description: AI Worker 回调此接口更新 ai_jobs 表状态。需要在 Gateway 层验证 Internal API Key。
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIJobStatusUpdate'
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
        "404":
          description: 任务不存在
...
  /internal/episodes/{episode_id}:
    get:
      tags: [AI Internal]
      summary: 查询剧集信息（内部接口）
      description: AI Worker 内部查询剧集元数据（含原始视频 URL）。
      parameters:
        - name: episode_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 剧集信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Episode'
        "404":
          description: 剧集不存在
...